// src/enums/FilledQuestionTypes.ts
export enum FilledQuestionTypes {
    FilledRequirementId = "filledRequirementId",
    QuestionId = "questionId",
    FulfillmentId = "fulfillmentId",
    InapplicableReasonId = "inapplicableReasonId",
    Comment = "comment",
    WorkStopped = "workStopped",
    ResolvedInPlace = "resolvedInPlace",
    PlannedResolveDate = "plannedResolveDate",
}